requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-Session==0.5.0
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.97.0
openai==1.3.0
python-dateutil==2.8.2